[#662] package MainView and SliderView
