+++ ios/CCC/UI/Slider/SliderView.swift
@@ -57,69 +57,3 @@ struct SliderView: View {
         }
     }
 }
-
-struct SlideView: View {
-    @Environment(\.colorScheme) var colorScheme
-
-    var title: String
-    var image: Image
-    var subTitle1: String
-    var subTitle2: String
-    var buttonText: String
-    var buttonAction: () -> Void
-
-    var body: some View {
-        ZStack {
-            Color(MR.colors().background_strong.get()).edgesIgnoringSafeArea(.all)
-            VStack {
-
-                Spacer()
-
-                Text(title)
-                    .multilineTextAlignment(.center)
-                    .font(.largeTitle)
-
-                image
-                    .frame(width: 196, height: 196, alignment: .center)
-                    .font(.system(size: 128))
-                    .accentColor(MR.colors().text.get())
-
-                Text(subTitle1)
-                    .lineSpacing(12)
-                    .multilineTextAlignment(.center)
-                    .font(.body)
-                    .padding(10)
-                    .padding(.bottom, 30)
-
-                Text(subTitle2)
-                    .lineSpacing(12)
-                    .multilineTextAlignment(.center)
-                    .padding(10)
-                    .font(.callout)
-
-                Spacer()
-
-                Divider()
-
-                HStack {
-                    Spacer()
-
-                    Button(
-                        action: buttonAction,
-                        label: {
-                            Text(buttonText)
-                                .font(.body)
-                                .foregroundColor(MR.colors().text.get())
-                        }
-                    ).padding(EdgeInsets(top: 10, leading: 10, bottom: 15, trailing: 15))
-                }
-            }
-        }
-    }
-}
-
-struct SliderViewPreviews: PreviewProvider {
-    static var previews: some View {
-        SliderView().colorScheme(.light)
-    }
-}
